sdk: move `sync` methods to `sync` module
